pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "physics_ai"
version = "0.1.0"
description = "Digital Twin for Unitree H1 Humanoid with MuJoCo MJX and JAX-native PPO"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = {text = "MIT"}
dependencies = [
    "mujoco>=3.1.0",
    "mujoco-mjx>=3.1.0",
    "jax>=0.4.25,<0.5.0",
    "jaxlib>=0.4.25,<0.5.0",
    "flax>=0.8.0,<0.9.0",
    "optax>=0.2.0,<0.3.0",
    "gymnasium>=0.29.0",
    "numpy>=1.26.0,<2.0.0",
    "pyyaml>=6.0",
    "tensorboard>=2.15.0",
    "chex>=0.1.8,<0.1.87",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
cuda = [
    "jax[cuda12_pip]>=0.4.25,<0.5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["physics_ai*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
